+++ FedData.Rcheck/tests/testthat.R
@@ -1,4 +1,0 @@
-library(testthat)
-library(FedData)
-
-test_check("FedData")