+++ tests/testthat/test.SSURGO.R
@@ -17,11 +17,11 @@ test_that("The SoilDB data queries work", {
   
   template = "CO670"
   q <- paste0("SELECT areasymbol, saverest FROM sacatalog WHERE areasymbol IN (",paste(paste0("'",template,"'"),collapse=','),");")
-  expect_is(SDA_query(q),"data.frame")
+  expect_is(FedData:::soils_query(q),"data.frame")
   
   template = "blah"
   q <- paste0("SELECT areasymbol, saverest FROM sacatalog WHERE areasymbol IN (",paste(paste0("'",template,"'"),collapse=','),");")
-  expect_error(SDA_query(q))
+  expect_error(FedData:::soils_query(q))
 })
 
 test_that("The SSURGO datasets are available at the correct URL", {
@@ -29,7 +29,7 @@ test_that("The SSURGO datasets are available at the correct URL", {
   
   template = "CO670"
   q <- paste0("SELECT areasymbol, saverest FROM sacatalog WHERE areasymbol IN (",paste(paste0("'",template,"'"),collapse=','),");")
-  q <- SDA_query(q)
+  q <- FedData:::soils_query(q)
   q$saverest <- as.Date(q$saverest,format="%m/%d/%Y")
   
   url <- paste("https://websoilsurvey.sc.egov.usda.gov/DSD/Download/Cache/SSA/wss_SSA_",q$areasymbol,"_[",q$saverest,"].zip",sep='')