+++ NEWS.md
@@ -1,5 +1,6 @@
 # FedData 3.0.0.9000
 * Updated ITRDB functions to remove data.table deps.
+* Changed `SDA_query` to `soils_query` to avoid namespace masking with SoilDB.
 
 # FedData 2.5.7
 * Removing many internet resource tests from CRAN, to satisfy: 'Packages which use Internet resources should fail gracefully with an informative message if the resource is not available (and not give a check warning nor error).'